refactor: Move private functions into utils file
